+++ src/app/entrance/student/dashboard/submission_status/_components/SubmissionStatus.tsx
@@ -1,0 +1,5 @@
+import { SubmissionStatusContents } from './SubmissionStatusContents'
+
+export default async function SubmissionStatus() {
+	return <SubmissionStatusContents />
+}
