+++ src/app/entrance/student/dashboard/top/_components/Top.tsx
@@ -1,0 +1,5 @@
+import { TopContents } from './TopContents'
+
+export default async function Top() {
+	return <TopContents />
+}
